Refactor usage of logging and fix import
